add defineImageFileHandlers
