+++ playground/components/common/upload-file.ts
@@ -1,4 +1,10 @@
-import type { Uploader } from 'prosekit/extensions/file'
+import { union, insertNode } from 'prosekit/core'
+import {
+  defineFileDropHandler,
+  defineFilePasteHandler,
+  UploadTask,
+  type Uploader,
+} from 'prosekit/extensions/file'
 
 /**
  * Uploads the given file to https://tmpfiles.org/ and returns the URL of the
@@ -7,7 +13,7 @@ import type { Uploader } from 'prosekit/extensions/file'
  * This function is only for demonstration purposes. All uploaded files will be
  * deleted after 1 hour.
  */
-export const tmpfilesUploader: Uploader<string> = ({
+const tmpfilesUploader: Uploader<string> = ({
   file,
   onProgress,
 }): Promise<string> => {
@@ -50,4 +56,44 @@ export const tmpfilesUploader: Uploader<string> = ({
     xhr.open('POST', 'https://tmpfiles.org/api/v1/upload', true)
     xhr.send(formData)
   })
+}
+
+/**
+ * Returns an extension that handles image file uploads when pasting or dropping
+ * images into the editor.
+ */
+export function defineImageFileHandlers() {
+  return union(
+    defineFilePasteHandler(({ view, file }) => {
+      if (!file.type.startsWith('image/')) {
+        return false
+      }
+      const uploadTask = new UploadTask({
+        file,
+        uploader: tmpfilesUploader,
+      })
+      // Insert the image node at the current text selection position
+      const command = insertNode({
+        type: 'image',
+        attrs: { src: uploadTask.objectURL },
+      })
+      return command(view.state, view.dispatch, view)
+    }),
+    defineFileDropHandler(({ view, file, pos }) => {
+      if (!file.type.startsWith('image/')) {
+        return false
+      }
+      const uploadTask = new UploadTask({
+        file,
+        uploader: tmpfilesUploader,
+      })
+      // Insert the image node at the drop position
+      const command = insertNode({
+        type: 'image',
+        attrs: { src: uploadTask.objectURL },
+        pos,
+      })
+      return command(view.state, view.dispatch, view)
+    }),
+  )
 }
