+++ playground/examples/react/full/extension.ts
@@ -1,14 +1,9 @@
 import { defineBasicExtension } from 'prosekit/basic'
-import { defineNodeAttr, insertNode, union } from 'prosekit/core'
+import { union } from 'prosekit/core'
 import {
   defineCodeBlock,
   defineCodeBlockShiki,
 } from 'prosekit/extensions/code-block'
-import {
-  defineFileDropHandler,
-  defineFilePasteHandler,
-  UploadTask,
-} from 'prosekit/extensions/file'
 import { defineHorizontalRule } from 'prosekit/extensions/horizontal-rule'
 import { defineMention } from 'prosekit/extensions/mention'
 import { definePlaceholder } from 'prosekit/extensions/placeholder'
@@ -19,7 +14,7 @@ import {
 
 import CodeBlockView from './code-block-view'
 import ImageView from './image-view'
-import { tmpfilesUploader } from './upload-file'
+import { defineImageFileHandlers } from './upload-file'
 
 export function defineExtension() {
   return union(
@@ -38,35 +33,7 @@ export function defineExtension() {
       name: 'image',
       component: ImageView satisfies ReactNodeViewComponent,
     }),
-    defineFilePasteHandler(({ view, file }) => {
-      if (!file.type.startsWith('image/')) {
-        return false
-      }
-      const uploadTask = new UploadTask({
-        file,
-        uploader: tmpfilesUploader,
-      })
-      const command = insertNode({
-        type: 'image',
-        attrs: { src: uploadTask.objectURL },
-      })
-      return command(view.state, view.dispatch, view)
-    }),
-    defineFileDropHandler(({ view, file, pos }) => {
-      if (!file.type.startsWith('image/')) {
-        return false
-      }
-      const uploadTask = new UploadTask({
-        file,
-        uploader: tmpfilesUploader,
-      })
-      const command = insertNode({
-        type: 'image',
-        attrs: { src: uploadTask.objectURL },
-        pos,
-      })
-      return command(view.state, view.dispatch, view)
-    }),
+    defineImageFileHandlers(),
   )
 }
