+++ playground/examples/react/image-view/extension.ts
@@ -7,6 +7,7 @@ import {
 } from 'prosekit/react'
 
 import ImageView from './image-view'
+import { defineImageFileHandlers } from './upload-file'
 
 export function defineExtension() {
   return union(
@@ -15,6 +16,7 @@ export function defineExtension() {
       name: 'image',
       component: ImageView satisfies ReactNodeViewComponent,
     }),
+    defineImageFileHandlers(),
   )
 }
