+++ playground/examples/react/image-view/upload-file.ts
@@ -1,0 +1,1 @@
+../../../components/common/upload-file.ts
