+++ playground/examples/vue/full/extension.ts
@@ -11,6 +11,7 @@ import { defineVueNodeView, type VueNodeViewComponent } from 'prosekit/vue'
 
 import CodeBlockView from './code-block-view.vue'
 import ImageView from './image-view.vue'
+import { defineImageFileHandlers } from './upload-file'
 
 export function defineExtension() {
   return union(
@@ -29,6 +30,7 @@ export function defineExtension() {
       name: 'image',
       component: ImageView as VueNodeViewComponent,
     }),
+    defineImageFileHandlers(),
   )
 }
