+++ playground/examples/vue/image-view/extension.ts
@@ -3,6 +3,7 @@ import { union } from 'prosekit/core'
 import { defineVueNodeView, type VueNodeViewComponent } from 'prosekit/vue'
 
 import ImageView from './image-view.vue'
+import { defineImageFileHandlers } from './upload-file'
 
 export function defineExtension() {
   return union(
@@ -11,6 +12,7 @@ export function defineExtension() {
       name: 'image',
       component: ImageView as VueNodeViewComponent,
     }),
+    defineImageFileHandlers(),
   )
 }
